Switch NPath to a struct and move (I)FileSystem.cs into it. Fix some bugs
While it's nice to have (I)FileSystem.cs separate, it's better if NiceIO.cs
is a self contained file, so that it's easy to update from upstream
and run all its tests, so this moves the interface and class into it.

This changes NPath into a struct, which requires a few adjustments to
the codebase, and I stumbled onto a couple of bugs in the process, so
fixed that as well.